+++ index.html
@@ -8,8 +8,9 @@
     <title>mindofjacob.ca | Portfolio | Front End Developer</title>
 </head>
 <body>
-    <img src="images/gallery-purple.png">
-    <section id="intro"></section>
+    <section id="intro">
+        <h1 class="hidden">Introduction</h1>
+    </section>
 <!--This is the template for the gallery shells-->
     <template id="portfolio-template">
         <section class="portfolio-pieces" data-key="">
@@ -18,18 +19,14 @@
         </section>
     </template>
 <!--This is the template for the modal that will open up when clicked on and display information about the project-->
-    <template id="modal-template">
-        <template class="modal-panel" data-key="">
-            <div class="company-logo"><img src=""></div>
-            <h2 class="m_name"></h2>
-            <h3 class="m_type"></h3>
-            <p class="m-desc"></p>
-        </section>
-    </template>
-
-
-
-
+<section id="modal-template">
+    <section class="modal-panel">
+        <div class="company-logo"><img src="slug"></div>
+        <h2 class="m_name"></h2>
+        <h3 class="m_type"></h3>
+        <p class="m-desc"></p>   
+    </section>
+</section> 
 <script src="js/main.js" type="module" defer></script>    
 </body>
 </html>
