+++ js/components/dataMiner.js
@@ -1,0 +1,10 @@
+function getData(){
+    fetch ("./includes/functions.php")
+    .then (res => res.json())
+    .then(data => {
+        console.table(data);
+    })
+    .catch(error => console.error(error));
+}
+
+export { getData }
