+++ js/main.js
@@ -1,17 +1,6 @@
-console.log("Hello");
-(() => {
-    console.log("Hello");
-    function getData(){
-        console.log("Hello");
-        fetch ("./includes/functions.php")
-        .then (res => res.json())
-        .then(data => {
-            console.table(data);
-        })
-        .catch(error => console.error(error));
-        
-    }
+import { getData } from "./components/dataMiner.js";
 
+(() => {
     getData();
